==> analysis/01_read_data.Rmd <==
---
title: "Read single-cell and meta-data"
author: "LazDaria"
date: "`r Sys.Date()`"
output: 
  workflowr::wflow_html
editor_options:
  chunk_output_type: inline
params: 
  input: "/mnt/Multimodal_Imaging_Daria/Publication/20240811_Zenodo-Upload"
  output: "/mnt/Multimodal_Imaging_Daria/Publication/20240811_R_output"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      error=FALSE,
                      messages=FALSE,
                      warning=FALSE)
```

# Read in the data {#read-data}

We  will first read in single-cell data, previously extracted from processed multi-modal (IMC,IF) images in [MapMetIP](https://github.com/TaschnerMandlGroup/MapMetIP) or downloaded from [Zenodo](https://zenodo.org/records/13220635/files/MapMetIP_ProcessedDataset.zip). For reading the data, we use the [imcRtools](https://github.com/BodenmillerGroup/imcRtools) package. This package uses the `read_steinbock` function to read data into a `SpatialExperiment` (SPE). The [SpatialExperiment](https://bioconductor.org/packages/release/bioc/html/SingleCellExperiment.html) class [@Righelli2021] is an extension of the `SingleCellExperiment` class. 
For more information, please refer to `?read_steinbock`. 

## Load libraries
<details>
   <summary>Load libraries</summary>
   
```{r load-libraries}
library(imcRtools)
library(stringr)
library(RColorBrewer)
library(circlize)
library(paletteer)
library(dittoSeq)
library(readxl)
library(writexl)
library(dplyr)
```
</details>

## Read single-cell measurement

Replace `path` with the path to the folder where you stored the single-cell data (e.g. `path/to/folder/MapMetIP_ProcessedDataset` in case processed data was downloaded from Zenodo). 
In [MapMetIP](https://github.com/TaschnerMandlGroup/MapMetIP), we exported single-cell marker measurements based on non-dilated and dilated (by 1px) segmentation masks to capture membrane signals. As both are required in the downstream analysis, we will read in both and combine them into one SPE object saving each expression matrix into one `assays`slot. Measurements based on non-dilated masks will be in the `assays`slot `counts` and the measurements based on dilated masks will be stored in `counts_1px`. 


```{r read-steinbock}
path <- file.path(params$input, "MapMetIP_ProcessedDataset")

spe_0px <- read_steinbock(path=path, image_file=NULL, panel_file = file.path(path, 'NB_Panel.csv'), intensities_folder = "intensities-0px",regionprops_folder = "regionprops", graphs_folder=NULL) 
spe_1px <- read_steinbock(path=path, image_file=NULL, panel_file = file.path(path, 'NB_Panel.csv'), intensities_folder = "intensities-1px",regionprops_folder = "regionprops", graphs_folder=NULL) 

#Combine SPE objects into one object
spe <- spe_0px
assay(spe, "counts_1px") <- assay(spe_1px, "counts")

#remove single SPE objects
rm(spe_0px)
rm(spe_1px)
```

## Read single-cell metadata

We set the `colnames` of the object to generate unique identifiers per cell.

```{r set-colnames}
colnames(spe) <- paste0(spe$sample_id, "_", spe$ObjectNumber)
```

When then read sample and patient-specific metadata from the metadata file provided on Zenodo [(Supplementary_Tables/20240808_Suppl_Table1.xlsx)](https://doi.org/10.5281/zenodo.13220634). 

```{r read-metadata}

file_path <- file.path(params$input, 'Supplementary_Tables/20240808_Suppl_Table1.xlsx')
sample_spec_metadata <- read_excel(file_path, sheet = "sample_spec_metadata")
patient_spec_clinical <- read_excel(file_path, sheet = "patient_spec_clinical")
patient_spec_genetic <- read_excel(file_path, sheet = "patient_spec_genetic")

# Merge patient and sample-specific data
merged_data <- sample_spec_metadata %>%
  left_join(patient_spec_clinical, by = "Patient_ID")

meta <- merged_data %>%
  left_join(patient_spec_genetic, by = "Patient_ID")
```

and assign it to the colData slot in our SPE object.

```{r add-coldata}

spe$sample <- gsub(spe$sample_id, pattern="([UM]).*", replacement="\\1")
spe$fm_id <- str_match(spe$sample, "_.*_(.*)_")[,2]  
spe$tissue <- str_match(spe$sample, "^.*_.*_(.*)")[,2] 
spe$tissue[spe$tissue=="TU"] = "PT"

spe$puncture_side <- meta$Puncture_side[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$imc_score <- meta$IMC_score[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$staining_date <- meta$Staining_date[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$stainer <- meta$Stainer[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$imc_date <- meta$IMC_date[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$control <- meta$control[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$patient_id <- meta$Patient_ID[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$timepoint <- meta$Timepoint[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$stage <- meta$INSS_stage[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$progression <- meta$efs[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$death <- meta$death[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$efs_dur <- meta$efs_dur[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$death_dur <- meta$death_dur[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$X1p_loss <- meta$X1p_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$X11q_loss <- meta$X11q_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$X17q_gain <- meta$X17q_gain[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$MYCN_amp <- meta$MYCN_amp[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$age_diag <- meta$age_at_diag[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$therapy <- meta$therapy[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]
spe$sex_aab <- meta$sex_aab[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample_ID, meta$Tissue))]

spe$ROI <- str_match(spe$sample_id, "^.*_.*_.*_.*_(.*)")[,2]
spe$year <- str_c("20", str_extract(spe$fm_id, "^[0-9]+"))

```

Before we save the SPE object, we check the distribution of markers.

```{r plot by staining date}

marker <- 'CD3_Sm152_mean'
tissue <- unique(spe$tissue)

dittoRidgePlot(spe[,spe$tissue%in%tissue], var = marker, group.by = "tissue", assay = "counts") +
    ggtitle(paste(marker,  "- per tissue"))

```

## Save object

Finally, the generated SPE object can be saved for further downstream processing and analysis.

```{r save-objects-read-data}

saveRDS(spe, file.path(params$output, "spe.rds"))
```

## Session Info

<details>
   <summary>SessionInfo</summary>
   
```{r, echo = FALSE}
sessionInfo()
```
</details>
