==> analysis/02_QC_1.Rmd <==
---
title: "Quality control - Part 1"
author: "LazDaria"
date: "`r Sys.Date()`"
output: 
  workflowr::wflow_html
editor_options:
  chunk_output_type: inline
params: 
  input: "/mnt/Multimodal_Imaging_Daria/Publication/20240811_Zenodo-Upload"
  output: "/mnt/Multimodal_Imaging_Daria/Publication/20240811_R_output"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      error=FALSE,
                      messages=FALSE,
                      warning=FALSE)
```

# Quality control - Exclusion of lost cells {#qc-1}
Segmentation in [MapMetIP](https://github.com/TaschnerMandlGroup/MapMetIP) is performed based on high-resolution IF images. However, due to the washing step between IF microscopy and IMC staining, some cells might be lost. These have to be excluded before downstream analysis.

## Load libraries
<details>
   <summary>Load libraries</summary>
   
```{r load-libraries}
library(ggplot2)
library(bruceR)
library(ComplexHeatmap)
library(viridis)
library(circlize)
library(RColorBrewer)
library(paletteer)
library(Rphenoannoy)
library(scuttle)
library(cytomapper)
library(cowplot)
library(gridGraphics)
```
</details>

## Read SPE object

We will first read the spatial experiment object (SPE) generated in the [read data section](#read-data). Please make sure, `path` points to the folder where you stored the SPE object.

```{r read-steinbock}

path <- params$output
spe <- readRDS(file.path(path,"spe.rds"))
```

## Exclude lost cells

### Quality control step 1
In a first step, we exclude cells, that have a mean intensity of zero for all IMC channels. For that, we will use the intensities measured based on the non-dilated masks (stored in `counts`) to avoid capturing signals from neighboring cells. To correctly identify lost cells in the clustering step below, we use the mean of the 20% highest pixel intensities as a measure for marker intensity here, as this feature is less dependent on cell size.

```{r cell-exclusion-1}

# Use measurements based on non-dilated mask stored in counts
intensity_mat <- assay(spe, "counts")
colnames(intensity_mat) <- spe$sample_id

#use mean-80 intensity to avoid losing big cells
intensity_mat_mean_80 <- intensity_mat[grepl("mean-80", rownames(intensity_mat)),]

#exclusion of all-zero cells
intensity_mat_imc <- intensity_mat_mean_80[!grepl("IF2_GD2|IF3_CD45|DAPI", rownames(intensity_mat_mean_80)),]
non_zeros <- as.logical(colSums(intensity_mat_imc)!=0)

spe <- spe[, non_zeros]
mat <- intensity_mat_mean_80[, non_zeros]
```

We excluded `r length(non_zeros[non_zeros==FALSE])` cells in this first quality control step.

For the next quality control step, we will only consider IMC and nuclear IF channels. 
```{r cell-exclusion-2}
mat <- mat[!grepl("IF2_GD2|IF3_CD45", rownames(mat)),]
```

### Quality control step 2
Although, we excluded all cells with zero mean intensities in all IMC channels, there might still be some cells, which actually swam off between IF and IMC and have IMC counts due to background noise or lateral spillover. In the heatmap below, we can clearly observe a cluster of cells expressing the IF marker DAPI,  but no IMC markers.

```{r data-visualization, fig.width=11, fig.height=9}
#named color vectors
qual_col_pals <- brewer.pal.info[brewer.pal.info$category == 'qual',]
col_vector_433 <- colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]

sample_names <- sub("_[^_]+$", "", colnames(mat))
col_sample <- setNames(sample(col_vector_433, length(unique(sample_names))), unique(sample_names))

set.seed(221228)
rand_cells <- sample(seq_len(ncol(mat)), 2000)

column_ha <- HeatmapAnnotation(sample = sample_names[rand_cells], 
                              col=list(sample=col_sample), 
                              show_legend=F)

Heatmap(mat[,rand_cells], 
        column_title = "Single-cell heatmap",
        top_annotation = column_ha, 
        show_column_dend = F,
        show_column_names=F,
        col=viridis(100)
        )
```

We will cluster the data to identify cells that swam off.

```{r phenograph-clustering}
set.seed(230619)

mat <- t(mat)
Rphenograph_out <- Rphenoannoy(mat, k = 45)

clusters <- factor(membership(Rphenograph_out[[2]]))

spe$pg_clusters_lostcells <- clusters
```

Next, we will have a look at the mean marker expression per cluster in a heatmap.

```{r data-visualization-2, fig.width=11, fig.height=9}
mean <- aggregate(mat, list(clusters), mean)
mean_t <- t(mean)
mean_t <- t(mean[,2:ncol(mean)])
colnames(mean_t) <- mean[,"Group.1"]

n <- length(unique(clusters))
col_clusters <- setNames(sample(col_vector_433, n), unique(clusters))

column_ha = HeatmapAnnotation(cluster = mean[,"Group.1"], 
                              ncells=anno_barplot(as.data.frame(table(clusters))[,"Freq"], height = unit(3, "cm")),
                              col=list(cluster=col_clusters), 
                              show_legend=T)

Heatmap(mean_t, 
        column_title = "Cluster-means heatmap",
        top_annotation = column_ha, 
        cluster_columns=T,
        show_column_dend = F,
        show_column_names=T,
        show_row_names=T,
        col=viridis(100)
        )
```

We can see that cluster 11 represents cells that swam off (DAPI expression without IMC marker expression).
To be sure that we only exclude cells that are really not present on the IMC channels, we will identify the samples with the highest number of cells from cluster 11.

```{r sanity-check}
c <- "11"
mat <- t(mat)

data <- data.frame(sample=colnames(mat), clusters)
data$freq <- 1

data_wide <- data %>% pivot_wider(names_from = c("clusters"), values_from = "freq", values_fn = sum, values_fill = 0)
data_long <- data_wide %>% pivot_longer(cols = !sample, names_to = "cluster", values_to = "count") 

clust <- data_long[data_long$cluster==c,]
cells_per_sample <- table(spe$sample_id)[clust$sample]
lost_cells_perc <- clust$count/cells_per_sample
clust$percentage <- lost_cells_perc
clust <- clust[order(clust$percentage, decreasing=T), ]
max_sample <- clust$sample[1]

write.csv(clust, file.path(params$output,"lost_cells.csv"))
```

Sample `r max_sample` has the highest number of cells from cluster 11. For a last sanity check, we will color the lost cells in the segmentation mask of that sample and plot the DAPI and the Iridium image, and the segmentation mask.

```{r sanity-check-2, fig.show='hide'}

masks <- loadImages(file.path(params$input, "MapMetIP_ProcessedDataset/masks", paste0(max_sample, ".tif")), as.is = TRUE)
img <- loadImages(file.path(params$input, "MapMetIP_ProcessedDataset/img", paste0(max_sample, ".tif")))

mcols(masks) <- DataFrame(sample_id = names(masks))
mcols(img) <- DataFrame(sample_id = names(img))
channelNames(img) <- rownames(spe)[1:44]

c = 11
tmp <- spe[,spe$sample_id == max_sample]
tmp$lost_cell <- "no"
tmp[,tmp$pg_clusters_lostcells==c]$lost_cell <- "yes"

dapi_plot <- plotPixels(img, 
                       colour_by = "IF1_DAPI_mean",
                       bcg = list(IF1_DAPI_mean = c(0, 1, 1)),
                       legend=NULL,
                       return_plot = TRUE)

iridium_plot <- plotPixels(img, 
                       colour_by = "DNA1_Ir193_mean",
                       bcg = list(DNA1_Ir193_mean = c(0, 1, 1)),
                       legend=NULL,
                       return_plot = TRUE)

mask_plot <- plotCells(masks,
                       object = tmp, 
                       cell_id = "ObjectNumber", 
                       img_id = "sample_id",
                       colour_by = "lost_cell",
                       colour = list(lost_cell = c("no" = "gray", "yes" = "red")),
                       missing_colour = "white",
                       legend=NULL,
                       return_plot = TRUE)
```
```{r plot-grid}
p1 <- ggdraw(dapi_plot$plot, clip = "on")
p2 <- ggdraw(iridium_plot$plot, clip = "on")
p3 <- ggdraw(mask_plot$plot, clip = "on")

plot_grid(p1, p2, p3)
```

By comparing the DAPI image, the Iridium image and the binary mask (highlighting cells from cluster 11 in blue), we see that cluster 11 definitely represents the cluster of lost cells. We excluded `r dim(spe[, spe$pg_clusters_lostcells==11])[2]` cells in total or `r dim(spe[, spe$pg_clusters_lostcells==11])[2]/649` cells per image (n=649).

## Save object

Finally, we exclude the lost cells from the SPE object and save the latter.

```{r save-objects-read-data}

spe <- spe[, spe$pg_clusters_lostcells!=11]
saveRDS(spe, file.path(params$output, "spe_qc1.rds"))
```

## Session Info

<details>
   <summary>SessionInfo</summary>
   
```{r, echo = FALSE}
sessionInfo()
```
</details>
