==> analysis/04_phenotyping.Rmd <==
---
title: "Cell phenotyping"
author: "LazDaria"
date: "`r Sys.Date()`"
output: 
  workflowr::wflow_html
editor_options:
  chunk_output_type: inline
params: 
  input: "/mnt/Multimodal_Imaging_Daria/Publication/20240811_Zenodo-Upload"
  output: "/mnt/Multimodal_Imaging_Daria/Publication/20240811_R_output"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      error=FALSE,
                      messages=FALSE,
                      warning=FALSE)
```

# Cell phenotyping {#phenotyping}
In this Rmd file, we will perform 
1. dimensionality reduction
2. phenograph clustering
3. annotate and merge clusters & metaclusters externally 
4. retrieve tumor cells from GD2-MPO double positive cells (caused by spillover) 
5. re-classify granulocyte metaclusters into maturity states based on morphology
6. define color codes for all celltypes and metaclusters

## Load libraries
<details>
   <summary>Load libraries</summary>
   
```{r load-libraries}
library(ComplexHeatmap)
library(RColorBrewer)
library(viridis)
library(scater)
library(BiocParallel)
library(Rphenoannoy)
library(tidyr)
library(tidyHeatmap)
library(bruceR)
library(dittoSeq)
library(dplyr)
library(mclust)
library(SingleR)
library(circlize)
library(readxl)
library(writexl)
```
</details>

## Read data
We will first read the data that passed the [second quality control step](#qc-2) and save the count data as a matrix. Low-quality markers (low SNR in a lot of samples) as well as all IF and nuclear IMC markers will be excluded for clustering. Moreover, mean intensity will be used to represent marker expression, to average out the signals arising from lateral spillover or artifacts. As explained in [the previous chapter](#qc2-save), we are using counts measured with respect to dilated masks in downstream analysis do capture membrane signal. 

```{r read-data}
spe <- readRDS(file.path(params$output,"spe_qc2.rds"))

intensity_mat <- assay(spe, "counts")
colnames(intensity_mat) <- colnames(spe)

#use mean intensity
intensity_mat <- intensity_mat[!grepl("mean-80", rownames(intensity_mat)),]

#exclude low-quality markers and functional marker cl.PARP
intensity_mat_hq <- intensity_mat[!grepl("PARP|IDO|DAPI|IF2_GD2|IF3_CD45_CD56|DNA1|H4K12Ac|H3K9Ac|CXCR2|PNMT|Fibro", rownames(intensity_mat)),]
intensity_mat_lq <- intensity_mat[grepl("PARP|IDO|DAPI|IF2_GD2|IF3_CD45_CD56|DNA1|H4K12Ac|H3K9Ac|CXCR2|PNMT|Fibro", rownames(intensity_mat)),]

non_zeros <- as.logical(colSums(intensity_mat_hq)!=0)

intensity_mat_hq_wo_zeros <- intensity_mat_hq[,non_zeros]
intensity_mat_lq_wo_zeros <- intensity_mat_lq[,non_zeros]

#Apply changes to SPE object
spe <- spe[,non_zeros]
rowData(spe)$use_channel <- !grepl("mean-80|PARP|IDO|DAPI|IF2_GD2|IF3_CD45_CD56|DNA1|H4K12Ac|H3K9Ac|CXCR2|PNMT|Fibro", rownames(spe))

##Matrix with morph. features
morph_mat_clipped <- (cbind("area"=spe$area_clipped, "solidity"=spe$solidity_clipped, "aspect_ratio"=spe$aspect_ratio_clipped))

tissue <- spe$tissue
cell_id <- colnames(spe)
```
We are excluding `r length(non_zeros[non_zeros==FALSE])` more cells here, as they are negative for all markers used in downstream analysis. 
To get a first grasp of the single-cell data and potential clusters, we will plot it in a single-cell heatmap.

```{r vis-scdata, fig.height=8}
set.seed(221228)

mat <- intensity_mat_hq_wo_zeros

qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]
col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]

col_sample = setNames(sample(col_vector_433, length(unique(spe$sample))), unique(spe$sample))

rand_cells <- sample(seq_len(ncol(mat)), 2000)

Heatmap(mat[,rand_cells], 
        column_title = "Single-cell heatmap",
        show_column_dend = F,
        show_column_names=F,
        col=viridis(100)
        )
```

## Dimensionality reduction prior to clustering {#umap1}
We reduce dimensionality of the data to five dimensions prior to clustering as described [here](https://umap-learn.readthedocs.io/en/latest/clustering.html). `calculateUMAP` is a wrapper function using the package `uwot`. `uwot` relies on the underlying compiler and C++ standard library on your machine and this can result in minor differences in output even with the same input data, arguments, packages and R version, as described [here](https://rdrr.io/github/jlmelville/uwot/f/vignettes/articles/reproducibility.Rmd). To ensure reproducibility of the exact same results between machines, we provide the dimensionality reduced count data, used for figure generation, as R object, on Zenodo [(R_intermediary/20240808_mat_red_clustering.rds)](https://doi.org/10.5281/zenodo.13220634). 

```{r UMAP-pre-clustering}
#mat_red_clustering <- calculateUMAP(mat, ncomponents=5, n_neighbors=30, min_dist= 0, BPPARAM = MulticoreParam(RNGseed = 221228))
mat_red_clustering <- readRDS(file.path(params$input, "R_intermediary/20230628_mat_red_clustering.rds")) 

rownames(mat_red_clustering) <- colnames(spe)
reducedDim(spe, "UMAP_clustering") <- mat_red_clustering
```

## Phenograph clustering
Next, we perform phenograph clustering on these five dimensions.

```{r phenograph-clustering}
set.seed(221228)

Rphenograph_out <- Rphenoannoy(mat_red_clustering, k = 45)
clusters <- factor(membership(Rphenograph_out[[2]]))

saveRDS(clusters, file.path(params$output, "pg_clusters.rds"))

spe$pg_clusters <- clusters
```

Let's plot the mean marker expressions per cluster in a bubble plot.

```{r vis-data-bubbleplot, fig.height=14, fig.width=16}
#Mean expression per cluster
exp_mat_ <- aggregate(t(mat), list(spe$pg_clusters), mean)
exp_mat <- exp_mat_[,2:ncol(exp_mat_)]
exp_mat <- apply(exp_mat, 2, RESCALE, to=0:1)
rownames(exp_mat) <- exp_mat_[,"Group.1"]
exp_mat <- t(exp_mat)

#Percentage of cells per cluster expressing a marker (>20th percentile)
tmp <- mat
tmp[tmp > unname(quantile(exp_mat[exp_mat>0], c(0.2)))] <- 1
tmp[tmp <= unname(quantile(exp_mat[exp_mat>0], c(0.2)))] <- 0
tmp <- aggregate(t(tmp), list(spe$pg_clusters), sum)
tmp2 <- tmp[,2:ncol(tmp)]
rownames(tmp2) <- tmp[,"Group.1"]
percent_mat <- tmp2 / unname(table(droplevels(spe$pg_clusters)))
percent_mat <- t(percent_mat)
percent_mat[percent_mat < 0.1] <- 0.1 #to make dots visible

#Annotation bars
bm_count <- table(spe[,spe$tissue=="BM"]$pg_clusters)
bm_count[setdiff(colnames(exp_mat), names(bm_count))] <- 0
bm_count <- bm_count[colnames(exp_mat)]
pt_count <- table(spe[,spe$tissue=="PT"]$pg_clusters)
pt_count[setdiff(colnames(exp_mat), names(pt_count))] <- 0
pt_count <- pt_count[colnames(exp_mat)]

mean_morph <- aggregate(morph_mat_clipped, list(spe$pg_clusters), mean)
area_mean = mean_morph[,"area"]
solidity_mean = mean_morph[,"solidity"]
col_mean_area = colorRamp2(c(min(area_mean),min(area_mean)+(max(area_mean)-min(area_mean))/2, max(area_mean)), c("blue", "white", "red"))
col_mean_solidity = colorRamp2(c(min(solidity_mean),min(solidity_mean)+(max(solidity_mean)-min(solidity_mean))/2, max(solidity_mean)), c("blue", "white", "red"))

col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]
col_clusters = setNames(sample(col_vector_433, length(unique(spe$pg_clusters))), unique(spe$pg_clusters))

column_ha = HeatmapAnnotation(pg_clusters = colnames(exp_mat),
                              roundness = solidity_mean,
                              ncells_PT=anno_barplot(unname(pt_count), add_numbers=F, height = unit(0.7, "cm"), gp = gpar(fill = "brown")),
                              ncells_bm=anno_barplot(unname(bm_count), add_numbers=F, height = unit(0.7, "cm"), gp = gpar(fill = "aquamarine4")),
                              col=list(pg_clusters=col_clusters,
                                       roundness = col_mean_solidity,
                                       area = col_mean_area), 
                              show_legend=T)

#Function to set dot size based on percentage expressed
cell_fun = function(j, i, x, y, w, h, fill){
          grid.rect(x = x, y = y, width = w, height = h, 
                    gp = gpar(col = NA, fill = NA))
          grid.circle(x=x,y=y,r= percent_mat[i, j]/2 * min(unit.c(w, h)),
                      gp = gpar(fill = col_fun(exp_mat[i, j]), col = NA))}

#Function to set dot color based on mean expression
col_fun = circlize::colorRamp2(c(min(exp_mat),(max(exp_mat)-min(exp_mat))/2, max(exp_mat)), viridis(100)[c(1,50,100)])

#Function to set legend size according to percentage expressed
layer_fun = function(j, i, x, y, w, h, fill){
          grid.rect(x = x, y = y, width = w, height = h, 
                    gp = gpar(col = NA, fill = NA))
          grid.circle(x=x,y=y,r= pindex(percent_mat, i, j)/1 * unit(1.5, "mm"),
                      gp = gpar(fill = col_fun(pindex(exp_mat, i, j)), col = NA))}

#Generate legend for percentage expressed
lgd_list = list(
    Legend( labels = c(0,0.25,0.5,0.75,1), title = "percentage expressed",
            graphics = list(
              function(x, y, w, h) grid.circle(x = x, y = y, r = 0 * unit(2, "mm"),
                                               gp = gpar(fill = "black")),
              function(x, y, w, h) grid.circle(x = x, y = y, r = 0.25 * unit(2, "mm"),
                                               gp = gpar(fill = "black")),
              function(x, y, w, h) grid.circle(x = x, y = y, r = 0.5 * unit(2, "mm"),
                                               gp = gpar(fill = "black")),
              function(x, y, w, h) grid.circle(x = x, y = y, r = 0.75 * unit(2, "mm"),
                                               gp = gpar(fill = "black")),
              function(x, y, w, h) grid.circle(x = x, y = y, r = 1 * unit(2, "mm"),
                                               gp = gpar(fill = "black")))
            ))

#Heatmap
hp <- Heatmap(exp_mat,
        heatmap_legend_param=list(title="expression"),
        column_title = "Bubble plot - normalized mean expression and %expressed", 
        top_annotation = column_ha, 
        col=col_fun,
        rect_gp = gpar(type = "none"),
        clustering_distance_columns="pearson", 
        clustering_method_columns="average", 
        layer_fun = layer_fun,
        row_labels=sapply(strsplit(rownames(exp_mat), "_"), "[[", 1),
        cell_fun = cell_fun,
        show_row_dend = F,
        column_dend_height = unit(1.5, "cm"),
        row_names_gp = gpar(fontsize = 10),
        column_names_gp = gpar(fontsize = 10),
        column_names_rot = 45,
        border = "black")

d1 <- draw( hp, annotation_legend_list = lgd_list)
```

For visualization purposes we will use different UMAP parameters. We read a previously generated R file from Zenodo [(R_intermediary/20240808_mat_red.rds)](https://doi.org/10.5281/zenodo.13220634) due to reasons explained in (#umap1).

```{r UMAP-visualization}
#mat_red <- calculateUMAP(mat, n_neighbors=30, BPPARAM = MulticoreParam(RNGseed = 221228))
mat_red <- readRDS(file.path(params$input, "R_intermediary/20230628_mat_red.rds")) 

rownames(mat_red) <- colnames(spe)
reducedDim(spe, "UMAP") <- mat_red
```

Let's visualize phenograph clusters and tissue types in UMAP space. Here `dittoDimPlot` can be used directly on the SPE object.

```{r UMAP-vis-clusters}
p1 <- dittoDimPlot(spe, var = "pg_clusters", 
             reduction.use = "UMAP", size = 1,
             do.label = TRUE,
             legend.show=F) +
  scale_color_manual(values = col_clusters) + 
  theme(legend.title = element_blank()) +
  ggtitle("PG clusters on UMAP")

p1
```

```{r UMAP-vis-marker, fig.width=10}
p1 <- dittoDimPlot(spe, var = "GD2_Gd155_mean", assay = "counts",
             reduction.use = "UMAP", size = 0.2, 
             colors = viridis(100), do.label = TRUE) +
    scale_color_viridis()

p2 <- dittoDimPlot(spe, var = "CD15_Bi209_mean", assay = "counts",
             reduction.use = "UMAP", size = 0.2, 
             colors = viridis(100), do.label = TRUE) +
    scale_color_viridis()

p1+p2
```

Next, we will plot the tissue type separately in the dim-red space.

```{r UMAP-vis-tissue}
var <- "tissue"

rownames(mat_red) <- spe$sample_id
meta <- colData(spe)[match(rownames(mat_red), spe$sample_id),var]

tissue_types <- unique(meta)
bin_col <- setNames(c(adjustcolor("gray90",alpha.f=1), "palegreen4"), c(FALSE,TRUE))

plot_list <- list()

for(i in 1: length(tissue_types)){

  sample_binary <- meta == tissue_types[i] 
  
  background <- as.data.frame(mat_red[!sample_binary,])
  foreground <- as.data.frame(mat_red[sample_binary,])
  plot_list[[i]] <- ggplot(background, aes(x=V1, y=V2) ) +
    geom_point(color="gray90", size=0.02) +
    geom_point(data=foreground, color="palegreen4", show.legend=F, size=0.02, alpha=0.25) +
    ggtitle(tissue_types[i]) +
    theme(plot.title = element_text(size = 5), 
          axis.title = element_text(size=5), 
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(), 
          panel.background = element_blank(), 
          axis.line = element_line(colour = "black"))
}


plot_list[[1]]
plot_list[[2]]
```

We can see that PT samples are primarily comprised of tumor cells (GD2+) and bone marrow samples include a large proprotion of granulocytes (CD15+).

## Annotation and merging of clusters

Next, we read assigned celltype and metacluster-labels for clusters from the metadata file provided on Zenodo [(Supplementary_Tables/20240808_Suppl_Table3.xlsx)](https://doi.org/10.5281/zenodo.13220634). These have been assigned based on thorough examination of images and co-expressions and using literature.
Let's start with celltype assignments...

```{r annotations-celltypes}
file_path <- file.path(params$input, 'Supplementary_Tables/20240808_Suppl_Table3.xlsx')
annotation <- read_excel(file_path, sheet = "annotation_v7")

BM_cluster_names <- setNames(annotation$BM_celltype, annotation$cluster)
PT_cluster_names <- setNames(annotation$PT_celltype, annotation$cluster)

BM_celltypes <- recode(spe[,spe$tissue=="BM"]$pg_clusters, !!!BM_cluster_names)
PT_celltypes <- recode(spe[,spe$tissue=="PT"]$pg_clusters, !!!PT_cluster_names)

spe$celltype <- "na"
spe[,spe$tissue=="BM"]$celltype<- as.vector(BM_celltypes)
spe[,spe$tissue=="PT"]$celltype<- as.vector(PT_celltypes)
```

...and then do the same for metaclusters:

```{r annotations-metaclusters}
BM_metacluster_names <- setNames(annotation$BM_metacluster, annotation$cluster)
PT_metacluster_names <- setNames(annotation$PT_metacluster, annotation$cluster)

BM_metacelltypes <- recode(spe[,spe$tissue=="BM"]$pg_clusters, !!!BM_metacluster_names)
PT_metacelltypes <- recode(spe[,spe$tissue=="PT"]$pg_clusters, !!!PT_metacluster_names)

spe$metacluster <- "na"
spe[,spe$tissue=="BM"]$metacluster<- as.vector(BM_metacelltypes)
spe[,spe$tissue=="PT"]$metacluster<- as.vector(PT_metacelltypes)
```

## Retrieve tumor cells from GD2/MPO double positive cells

There is one cluster in BM samples, which expresses both GD2 and MPO. This is caused by lateral spillover in samples densely populated with either tumor cells (GD2+) or granulocytes (MPO+). We will retrieve disseminated tumor cells (DTCs) from this cluster using GMMs.

```{r GD2-MPO-subclustering, fig.height=6, fig.width=6}
set.seed(20231201)

markers <- grepl("GD2|CD45|MPO|CD15|CD11b", rownames(spe)) & !grepl("mean-80|IF", rownames(spe))

tmp <- t(assay(spe[,spe$celltype=="GD2+MPO+"], "counts")[markers,])

gmm <- Mclust(tmp, G = 2)
clusters <- gmm$classification

rand_cells <- sample(seq_len(nrow(tmp)), 500)

column_ha = HeatmapAnnotation(cluster = clusters[rand_cells], 
                              show_legend=T)

tmp2 <- t(assay(spe[,spe$celltype=="GD2+MPO+"], "counts")[rowData(spe)$use_channel,])

ht<- Heatmap(t(tmp2[rand_cells,]), 
        column_title = "Single-cell heatmap",
        cluster_columns=F,
        show_row_dend = F,
        row_labels=sapply(strsplit(colnames(tmp2[rand_cells,]), "_"), "[[", 1),
        top_annotation=column_ha,
        column_order= order(unname(clusters[rand_cells]), decreasing=F),
        show_column_dend = T,
        show_column_names=F,
        col=viridis(100)
        )

d1 <- draw(ht)
```

In the heatmap above, we can see a clear separation of CD45+ MPO+ CD15+ CD11b+ (cluster 1) from GD2+ GD2+ CD45- MPOlow CD15- CD11b- (cluster 2) cells. After thorough investigation on the image-level, cluster 2 was assigned to DTCs and cluster 1, a mix of DTCs and granulocytes, to “other”. 

Finally, the tumor subphenotypes of DTCs within GMM cluster 2 is determined based on similarity in expression profiles to the identified tumor subtypes from above, using singleR.

```{r singleR-tumor-cell-assignment}
markers <- rowData(spe)$use_channel

query <- t(tmp2[clusters==2,colnames(tmp2)%in%rownames(spe)[markers]])
ref <- assay(spe[markers,spe$metacluster=="tumor"], "counts")
ref_clusters <- spe[markers,spe$metacluster=="tumor"]$celltype

predicted_labels <- SingleR(test = query, ref = ref, labels = ref_clusters, BPPARAM = MulticoreParam(workers=10, RNGseed=20231030))

col_clusters <- c("CD24+ marker-lo TC"=rgb(230,245,201, maxColorValue = 255),
                "early NB TC"=rgb(204,235,197, maxColorValue = 255),
                "bridge-like TC"=rgb(198,255,72, maxColorValue = 255),
                "GATA3hi TC"=rgb(65,208,57, maxColorValue = 255),
                "CD24- marker-lo TC"=rgb(60,119,97, maxColorValue = 255),
                "CHGAhi TC"=rgb(102,166,30, maxColorValue = 255),
                "Ki67hi TC"=rgb(149,144,89, maxColorValue = 255),
                "CXCR4hi TC"=rgb(210,205,126, maxColorValue = 255),
                "CD44+ TC"=rgb(27,158,119, maxColorValue = 255),
                "GD2lo TC"=rgb(102,194,165, maxColorValue = 255))

plotScoreHeatmap(predicted_labels, annotation_colors=list(Labels = col_clusters), annotation_legend=F, treeheight_row=0) #define the colors first 

labels <- predicted_labels$labels

#Assign predicted labels
dtc <- rownames(tmp2[clusters==2,])
no_dtc <- rownames(tmp2[clusters==1,])
spe[,colnames(spe)%in%dtc]$celltype <- labels
spe[,colnames(spe)%in%no_dtc]$celltype <- "other"
spe[,colnames(spe)%in%dtc]$metacluster <- "tumor"
```
We can see that most cells were assigned to either CD24+ or CD24- marker-lo TCs, which is reasonable based on their expression.

## Granulocyte re-classification

For BM samples, granulocytes are re-classified into four maturation states based on the morphological feature solidity.

```{r granulocyte_subclustering}
spe_gran <- spe[,spe$celltype=="granulocyte" & spe$tissue=="BM"]

spe_gran$gran_type <- 'neutrophil'
spe_gran[,spe_gran$solidity>=0.83 & spe_gran$solidity<0.93]$gran_type <- 'band cell'
spe_gran[,spe_gran$solidity>=0.93 & spe_gran$solidity<0.97]$gran_type <- 'myelocyte'
spe_gran[,spe_gran$solidity>=0.97]$gran_type <- 'myeloblast'

#assign new granulocyte celltypes
spe[,colnames(spe_gran)]$celltype <- spe_gran$gran_type
spe[,spe$tissue=="PT" & spe$celltype=="granulocyte"]$celltype <- "neutrophil"
```

## Assign colors 
Finally, we assign colors to all celltypes and metaclusters.

```{r colors-celltypes}
col_clusters <- c("other"=rgb(255,255,209, maxColorValue = 255),
                "LIN- Ki67+"=rgb(254,217,166, maxColorValue = 255),
                "HSPC"=rgb(253,180,98, maxColorValue = 255),
                "GZMB+ S100B- DC/NK"=rgb(28,146,255, maxColorValue = 255),
                "B cell"=rgb(168,128,250, maxColorValue = 255),
                "GZMB+ S100B+ DC/NK"=rgb(0,255,255, maxColorValue = 255),
                "CD8+ naive T cell"=rgb(188,128,189, maxColorValue = 255),
                "CD3+ GZMB+ T cell"=rgb(190,174,212, maxColorValue = 255),
                "CD8+ S100B+ T cell"=rgb(253,218,236, maxColorValue = 255),
                "CD8+ GZMB+ PD1lo T cell"=rgb(251,154,153, maxColorValue = 255),
                "CD4+ naive T cell"=rgb(242,3,137, maxColorValue = 255),
                "CD3+ T cell"=rgb(251,128,114, maxColorValue = 255),
                "CD4+ PD1+ T cell"=rgb(254,142,175, maxColorValue = 255),
                "CD8+ PDL1lo T cell"=rgb(240,0,240, maxColorValue = 255),
                "CD4+ S100B+ T cell"=rgb(227,26,28, maxColorValue = 255),
                "dense T cell region"=rgb(145,2,144, maxColorValue = 255),
                "CD24+ marker-lo TC"=rgb(230,245,201, maxColorValue = 255),
                "early NB TC"=rgb(204,235,197, maxColorValue = 255),
                "neural progenitor"=rgb(238,118,0, maxColorValue = 255),
                "bridge-like TC"=rgb(198,255,72, maxColorValue = 255),
                "GATA3hi TC"=rgb(65,208,57, maxColorValue = 255),
                "CD24- marker-lo TC"=rgb(60,119,97, maxColorValue = 255),
                "CD14+ PDL1- MO"=rgb(104,180,238, maxColorValue = 255),
                "monoblast-like"=rgb(128,177,211, maxColorValue = 255),
                "CD14+ PDL1+ MO"=rgb(51,180,180, maxColorValue = 255),
                "CHGAhi TC"=rgb(102,166,30, maxColorValue = 255),
                "Ki67hi TC"=rgb(149,144,89, maxColorValue = 255),
                "CXCR4hi TC"=rgb(210,205,126, maxColorValue = 255),
                "CD44+ TC"=rgb(27,158,119, maxColorValue = 255),
                "GD2lo TC"=rgb(102,194,165, maxColorValue = 255),
                "CD14- PDL1- MO/DC"=rgb(39,130,187, maxColorValue = 255),
                "CD14- PDL1+ MO/DC"=rgb(166,206,227, maxColorValue = 255),
                "pDC"=rgb(212,241,240, maxColorValue = 255),
                "band cell"=rgb(169,189,210, maxColorValue = 255),
                "myelocyte"=rgb(217,217,217, maxColorValue = 255),          
                "myeloblast"=rgb(2,0,138, maxColorValue = 255),
                "neutrophil"=rgb(113,113,113, maxColorValue = 255),
                "fibroblast/endothel"=rgb(212,89,67, maxColorValue = 255),
                "schwann cell"=rgb(156,76,24, maxColorValue = 255),
                "MSC"=rgb(174,129,37, maxColorValue = 255),
                "HPC"=rgb(212,143,75, maxColorValue = 255),
                "Ki67+ CXCR4+ cell"=rgb(254,215,0, maxColorValue = 255))

metadata(spe)$color_vectors$col_celltype <- col_clusters
```

```{r colors-metaclusters}
col_metacluster <- c("tumor"="palegreen4",
                  "T cell"="indianred",
                  "granulocyte"="#80B1D3",
                  "MO/DC/NK"="#8DD3C7",
                  "B cell"="#BC80BD",
                  "other"="goldenrod1",
                  "progenitor"="tan2",
                  "mesenchymal"="tan4")
                
metadata(spe)$color_vectors$col_metacluster <- col_metacluster
```

## Visualizations
For visualization purpose, we will give the celltypes a meaningful order. This order can be downloaded from a csv file from Zenodo [(R_intermediary/celltype_order.csv)](https://doi.org/10.5281/zenodo.13220634)

```{r order-celltypes}
clusternames_ordered <- read.csv(file.path(params$input, "R_intermediary/celltype_order.csv"), sep=";", colClasses = c("order" = "character"))
codes <- setNames(paste0(clusternames_ordered$order, "_", clusternames_ordered$celltype), clusternames_ordered$celltype)
spe$celltype_ordered <- recode(spe$celltype, !!!codes)

metadata(spe)$color_vectors$col_celltype_ordered <- metadata(spe)$color_vectors$col_celltype
names(metadata(spe)$color_vectors$col_celltype_ordered) <- recode(names(metadata(spe)$color_vectors$col_celltype), !!!codes)
```

As a last step, we generate UMAP and bubble plots for the final celltypes/metaclusters.

```{r umap-final-celltypes, fig.width=10}
p1 <- dittoDimPlot(spe, var = "celltype_ordered", 
             reduction.use = "UMAP", size = 0.3, 
             do.label = F,
             order = "randomize",
             legend.show=T) +
  scale_color_manual(values = metadata(spe)$color_vectors$col_celltype_ordered) +
  theme(legend.title = element_blank()) +
  ggtitle("Celltypes on UMAP")

p2 <- dittoDimPlot(spe, var = "metacluster", 
             reduction.use = "UMAP", size = 0.3, labels.size=4,
             do.label = T,
             order = "randomize",
             legend.show=F) +
  scale_color_manual(values = metadata(spe)$color_vectors$col_metacluster) +
  theme(legend.title = element_blank()) +
  ggtitle("Metaclusters on UMAP")

p1
```

```{r vis-metacluster}
p2
```


```{r bubbleplot-final-celltypes, fig.height=10, fig.width=12}
exp_mat_ <- aggregate(t(mat), list(spe$celltype), mean)
exp_mat <- exp_mat_[,2:ncol(exp_mat_)]
exp_mat <- apply(exp_mat, 2, RESCALE, to=0:1)
rownames(exp_mat) <- exp_mat_[,"Group.1"]
exp_mat <- t(exp_mat)

#Percentage of cells per cluster expressing a marker (>0)
tmp <- mat
tmp[tmp > unname(quantile(exp_mat[exp_mat>0], c(0.2)))] <- 1
tmp[tmp <= unname(quantile(exp_mat[exp_mat>0], c(0.2)))] <- 0
tmp <- aggregate(t(tmp), list(spe$celltype), sum)
tmp2 <- tmp[,2:ncol(tmp)]
rownames(tmp2) <- tmp[,"Group.1"]
percent_mat <- tmp2 / unname(table(spe$celltype))
percent_mat <- t(percent_mat)
percent_mat[percent_mat < 0.1] <- 0.1 #for visualization

#Annotation bars
bm_count <- table(spe[,spe$tissue=="BM"]$celltype)
bm_count[setdiff(colnames(exp_mat), names(bm_count))] <- 0
bm_count <- bm_count[colnames(exp_mat)]
pt_count <- table(spe[,spe$tissue=="PT"]$celltype)
pt_count[setdiff(colnames(exp_mat), names(pt_count))] <- 0
pt_count <- pt_count[colnames(exp_mat)]

mean_metaclust <-aggregate(spe$metacluster, list(spe$celltype), max)

mean_morph <- aggregate(morph_mat_clipped, list(spe$celltype), mean)
solidity_mean = mean_morph[,"solidity"]
col_mean_solidity = colorRamp2(c(min(solidity_mean),min(solidity_mean)+(max(solidity_mean)-min(solidity_mean))/2, max(solidity_mean)), c("blue", "white", "red"))

column_ha = HeatmapAnnotation(metacluster=mean_metaclust$x,
                              roundness = solidity_mean,
                              ncells_PT=anno_barplot(unname(pt_count), add_numbers=F, height = unit(0.7, "cm"), gp = gpar(fill = "brown")),
                              ncells_bm=anno_barplot(unname(bm_count), add_numbers=F, height = unit(0.7, "cm"), gp = gpar(fill = "aquamarine4")),
                              col=list(metacluster=metadata(spe)$color_vectors$col_metacluster, 
                                       roundness = col_mean_solidity
                                       ), 
                              show_legend=T)

#Function to set dot size based on percentage expressed
cell_fun = function(j, i, x, y, w, h, fill){
          grid.rect(x = x, y = y, width = w, height = h, 
                    gp = gpar(col = NA, fill = NA))
          grid.circle(x=x,y=y,r= percent_mat[i, j]/2 * min(unit.c(w, h)),
                      gp = gpar(fill = col_fun(exp_mat[i, j]), col = NA))}

#Function to set dot color based on mean expression
col_fun = circlize::colorRamp2(c(min(exp_mat),(max(exp_mat)-min(exp_mat))/2, max(exp_mat)), viridis(100)[c(1,50,100)])

#Function to set legend size according to percentage expressed
layer_fun = function(j, i, x, y, w, h, fill){
          grid.rect(x = x, y = y, width = w, height = h, 
                    gp = gpar(col = NA, fill = NA))
          grid.circle(x=x,y=y,r= pindex(percent_mat, i, j)/1 * unit(1.5, "mm"),
                      gp = gpar(fill = col_fun(pindex(exp_mat, i, j)), col = NA))}

#Generate legend for percentage expressed
lgd_list = list(
    Legend( labels = c(0,0.25,0.5,0.75,1), title = "percentage expressed",
            graphics = list(
              function(x, y, w, h) grid.circle(x = x, y = y, r = 0 * unit(2, "mm"),
                                               gp = gpar(fill = "black")),
              function(x, y, w, h) grid.circle(x = x, y = y, r = 0.25 * unit(2, "mm"),
                                               gp = gpar(fill = "black")),
              function(x, y, w, h) grid.circle(x = x, y = y, r = 0.5 * unit(2, "mm"),
                                               gp = gpar(fill = "black")),
              function(x, y, w, h) grid.circle(x = x, y = y, r = 0.75 * unit(2, "mm"),
                                               gp = gpar(fill = "black")),
              function(x, y, w, h) grid.circle(x = x, y = y, r = 1 * unit(2, "mm"),
                                               gp = gpar(fill = "black")))
            ))

#Heatmap
hp <- Heatmap(exp_mat,
        heatmap_legend_param=list(title="expression"),
        column_title = "Bubble plot - normalized mean expression and %expressed", 
        top_annotation = column_ha, 
        col=col_fun,
        rect_gp = gpar(type = "none"),
        clustering_distance_columns="pearson", 
        clustering_method_columns="average", 
        layer_fun = layer_fun,
        row_labels=sapply(strsplit(rownames(exp_mat), "_"), "[[", 1),
        cell_fun = cell_fun,
        show_row_dend = F,
        column_dend_height = unit(1.5, "cm"),
        row_names_gp = gpar(fontsize = 10),
        column_names_gp = gpar(fontsize = 10),
        column_names_rot = 45,
        border = "black")

d1 <- draw( hp, annotation_legend_list = lgd_list)
```

## Save object

```{r save-objects}

saveRDS(spe, file.path(params$output, "spe_final.rds"))
```

## Session Info

<details>
   <summary>SessionInfo</summary>
   
```{r, echo = FALSE}
sessionInfo()
```
</details>
