==> analysis/11_cytokine_assays.Rmd <==
---
title: "Heatmap - Cytokine Assays"
author: "LazDaria"
date: "`r Sys.Date()`"
output: 
  workflowr::wflow_html
editor_options:
  chunk_output_type: inline
params: 
  input: "/mnt/Multimodal_Imaging_Daria/Publication/20240811_Zenodo-Upload"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      error=FALSE,
                      messages=FALSE,
                      warning=FALSE)
```

# Heatmap for cytokine assays {#cyto}

In this Rmd file, we will plot a heatmap to visualize the abundance of cytokines in co-cultivation experiments. 

# Load libraries
<details>
   <summary>Load libraries</summary>
   
```{r load-libraries}
library(readxl)
library(dplyr)
library(stringr)
library(tidyr)
library(ComplexHeatmap)
library(viridis)
library(openxlsx)
library(bruceR)
```
</details>

## Read data
First, we load normalized data from three cytokine assays from Supplementary Table 4 provided on Zenodo [(Supplementary_Tables/20240713_Suppl_Table4.xlsx)](https://doi.org/10.5281/zenodo.13220634). 

```{r read-data, message=FALSE}
data_inf <- read_excel(file.path(params$input, 'Supplementary_Tables/20240713_Suppl_Table4.xlsx'), sheet = "INF3")
data_che <- read_excel(file.path(params$input, 'Supplementary_Tables/20240713_Suppl_Table4.xlsx'), sheet = "CHE1")
data_rec <- read_excel(file.path(params$input, 'Supplementary_Tables/20240713_Suppl_Table4.xlsx'), sheet = "REC1")

#Remove positive controls
data_inf <- data_inf[!grepl("POS", data_inf$Normalization),]
data_che <- data_che[!grepl("POS", data_che$Normalization),]
data_rec <- data_rec[!grepl("POS", data_rec$Normalization),]

head(data_che)
```

## Rename column names

```{r rename-col}
# Rename columns: remove dots and numbers, replace spaces with underscores
clean_colnames <- function(colname) {
  colname %>%
    str_replace_all("\\.\\d*$", "") %>%  # Remove trailing numbers after a dot
    str_replace_all("\\.\\.\\.+\\d*", "") %>%  # Remove dots and trailing numbers
    str_replace_all(" ", "_")                  # Replace spaces with underscores
}

# Apply the renaming function
colnames(data_inf) <- sapply(colnames(data_inf), clean_colnames)
colnames(data_che) <- sapply(colnames(data_che), clean_colnames)
colnames(data_rec) <- sapply(colnames(data_rec), clean_colnames)
```

## Compute mean between duplicate rows

```{r mean-duplicate-rows}
# Combine duplicate columns by calculating the mean
combine_duplicate_columns <- function(df) {
  df %>%
     pivot_longer(-Normalization, names_to = "colname", values_to = "value") %>%
     group_by(Normalization, colname) %>%
     summarise(mean_value = mean(value, na.rm = TRUE), .groups = 'drop') %>%
     pivot_wider(names_from = colname, values_from = mean_value)
}

# Apply the function to combine duplicate columns
data_combined_inf <- combine_duplicate_columns(data_inf)
data_combined_che <- combine_duplicate_columns(data_che)
data_combined_rec <- combine_duplicate_columns(data_rec)

# Remove dots from the column names
colnames(data_combined_inf) <- str_replace_all(colnames(data_combined_inf), "\\.", "")
colnames(data_combined_che) <- str_replace_all(colnames(data_combined_che), "\\.", "")
colnames(data_combined_rec) <- str_replace_all(colnames(data_combined_rec), "\\.", "")

#Set first column as rownames
data_combined_inf <- as.data.frame(data_combined_inf)
rownames(data_combined_inf) <- data_combined_inf$Normalization
data_combined_inf <- data_combined_inf[-1]

data_combined_che <- as.data.frame(data_combined_che)
rownames(data_combined_che) <- data_combined_che$Normalization
data_combined_che <- data_combined_che[-1]

data_combined_rec <- as.data.frame(data_combined_rec)
rownames(data_combined_rec) <- data_combined_rec$Normalization
data_combined_rec <- data_combined_rec[-1]
```

## Calculate log2FC 

We will then calculate the log2FC between mono- and co-cultures and filter those that are increasing upon co-culture in nMNA, MNA or both.

```{r log2FC}
data_all <- rbind(data_combined_inf, data_combined_che, data_combined_rec)
df <- as.data.frame(data_all)
col_filter_chla <- c("I", "PBMC_I", "CHLA", "CHLA_D")
col_filter_clb <- c("R", "PBMC_R", "CLB", "CLB_D")

df_unscaled <- cbind(df[,col_filter_chla], df[,col_filter_clb])

# Create the new columns with the specified calculations
df_unscaled <- df_unscaled %>%
  mutate(
    log2FC_CHLA_D_I = log2(CHLA_D / I),
    log2FC_CLB_D_R = log2(CLB_D / R),
    log2FC_CHLA_D_PBMC_I = log2(CHLA_D / PBMC_I),
    log2FC_CLB_D_PBMC_R = log2(CLB_D / PBMC_R),
    log2FC_CHLA_D_CHLA = log2(CHLA_D / CHLA),
    log2FC_CLB_D_CLB = log2(CLB_D / CLB)
  )

# Add the new column based on the condition
df_unscaled <- df_unscaled %>%
  mutate(both = if_else(
    log2FC_CHLA_D_I >= 0.3 & 
    log2FC_CLB_D_R >= 0.3 & 
    log2FC_CHLA_D_PBMC_I >= 0.3 & 
    log2FC_CLB_D_PBMC_R >= 0.3 & 
    log2FC_CHLA_D_CHLA >= 0.3 & 
    log2FC_CLB_D_CLB >= 0.3, 
    1, 
    0
  )) %>%
    mutate(nMNA = if_else(
    log2FC_CHLA_D_I >= 0.3 & 
    log2FC_CHLA_D_PBMC_I >= 0.3 & 
    log2FC_CHLA_D_CHLA >= 0.3 & 
    both == 0, 
    1, 
    0
  )) %>%
    mutate(MNA = if_else(
    log2FC_CLB_D_R >= 0.3 & 
    log2FC_CLB_D_PBMC_R >= 0.3 & 
    log2FC_CLB_D_CLB >= 0.3 &
    both == 0, 
    1, 
    0
  ))

df_up <- df_unscaled %>% filter(both == 1 | nMNA == 1 | MNA == 1)
```

## Create heatmap 
Let's plot a heatmap of the cytokines enriched in nMNA only:

```{r heatmap-nMNA, fig.width=19, fig.height=7}
df <- df_up

#Annotation
col_vector <- c("0"="#ffffbf", "1"="#fc8d59")
column_ha <- HeatmapAnnotation(nMNA = df_up$nMNA,
                               MNA = df_up$MNA,
                               both = df_up$both,
                               col=list(nMNA=col_vector,
                                        MNA=col_vector,
                                        both=col_vector))

#Separate scaling per cell line
col_filter_chla <- c("I", "PBMC_I", "CHLA", "CHLA_D")
col_filter_clb <- c("R", "PBMC_R", "CLB", "CLB_D")
data_sep_scale <- rbind(scale(t(df_up[,col_filter_chla])), scale(t(df_up[,col_filter_clb])))

# Define a diverging color palette
diverging_palette <- coolwarm_hcl <- colorspace::diverging_hcl(11, h = c(250, 10), c = 100, l = c(37, 88), power = c(0.7, 1.7))

Heatmap(data_sep_scale[, df_up$nMNA==1], 
        column_title = "Cytokine assays- CHLA90 and CLB-MA",
        row_order = c("I", "R", "PBMC_I", "PBMC_R", "CHLA", "CLB", "CHLA_D", "CLB_D"),
        cluster_columns =T,
        col=diverging_palette
        )
```

in MNA only:

```{r heatmap-MNA, fig.width=19, fig.height=7}
Heatmap(data_sep_scale[, df_up$MNA==1], 
        column_title = "Cytokine assays- CHLA90 and CLB-MA",
        row_order = c("I", "R", "PBMC_I", "PBMC_R", "CHLA", "CLB", "CHLA_D", "CLB_D"),
        cluster_columns =T,
        col=diverging_palette
        )
```

or in both:
```{r heatmap-both, fig.width=19, fig.height=7}
Heatmap(data_sep_scale[, df_up$both==1], 
        column_title = "Cytokine assays- CHLA90 and CLB-MA",
        row_order = c("I", "R", "PBMC_I", "PBMC_R", "CHLA", "CLB", "CHLA_D", "CLB_D"),
        cluster_columns =T,
        col=diverging_palette
        )
```

## Session Info

<details>
   <summary>SessionInfo</summary>
   
```{r, echo = FALSE}
sessionInfo()
```
</details>
